backend/requirements.txt
# FastAPI 核心
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.49
asyncpg
alembic==1.13.1

# 认证
python-jose[cryptography]>=3.3.0,<4
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Pydantic
pydantic==2.13.3
pydantic-settings==2.14.0

# 工具
python-dotenv==1.0.0
httpx==0.26.0
Pillow==12.1.0
python-docx==1.1.0
openpyxl==3.1.2
aiofiles==23.2.1
pdf2image==1.17.0

# 日期处理
python-dateutil==2.8.2
